requirements.txt
alembic==1.7.5
anyio==3.5.0
asgiref==3.4.1
black==21.12b0
click==8.0.3
fastapi==0.72.0
greenlet==1.1.2
h11==0.12.0
idna==3.3
importlib-metadata==4.10.1
importlib-resources==5.4.0
Mako==1.1.6
MarkupSafe==2.0.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.1
psycopg2-binary==2.9.3
pydantic==1.9.0
python-dotenv==0.19.2
sniffio==1.2.0
SQLAlchemy==1.4.29
starlette==0.17.1
tomli==1.2.3
typing_extensions==4.0.1
uvicorn==0.17.0
zipp==3.7.0
